update to diff_count; mergedata added to repository
`diff_count` modified so that some of the events (ions moving and
Frenkel pair creation) have been better encapsulated as methods of site
instances

`mergedata` added to repository. This utility takes a list of .xy
format files and merges them according to column 1, either giving the
average ( `-a` ) or sum ( `-s` ) of corresponding entries in column 2.