mergedata updated
`mergedata` modified to handle data sets starting at zero.
The code for outputting the merged data has been changed so that it is
now performed by the same code for all options, where the number to be
outputted is first passed to a proc, which is now selected by the
choice of output options.